chore: deduplicate time parsing
