+++ src/sync.rs
@@ -26,6 +26,7 @@ use crate::experiments::api::create_experiment;
 use crate::http::ApiClient;
 use crate::projects::api::{create_project, list_projects, Project};
 use crate::ui::{animations_enabled, fuzzy_select, is_quiet};
+use crate::utils::parse_duration_to_seconds;
 
 const STATE_SCHEMA_VERSION: u32 = 1;
 const DEFAULT_PULL_LIMIT: usize = 100;
@@ -2603,34 +2604,6 @@ fn build_root_spans_query(
     parts.join(" | ")
 }
 
-fn parse_duration_to_seconds(input: &str) -> Result<u64> {
-    let trimmed = input.trim();
-    if trimmed.is_empty() {
-        bail!("duration cannot be empty");
-    }
-    if let Ok(seconds) = trimmed.parse::<u64>() {
-        return Ok(seconds);
-    }
-
-    let suffix = trimmed.chars().last().filter(|ch| ch.is_ascii_alphabetic());
-    let (num_str, unit) = match suffix {
-        Some(unit) => (&trimmed[..trimmed.len() - unit.len_utf8()], unit),
-        None => (trimmed, 's'),
-    };
-    let value: u64 = num_str
-        .trim()
-        .parse()
-        .with_context(|| format!("invalid duration '{input}'"))?;
-    let multiplier = match unit.to_ascii_lowercase() {
-        's' => 1,
-        'm' => 60,
-        'h' => 60 * 60,
-        'd' => 60 * 60 * 24,
-        _ => bail!("invalid duration '{input}'. expected suffix s/m/h/d"),
-    };
-    Ok(value.saturating_mul(multiplier))
-}
-
 fn build_time_filter_clause(window: &str, extra_filter: Option<&str>) -> Result<String> {
     let seconds = parse_duration_to_seconds(window)?;
     let time_clause = format!("created >= NOW() - INTERVAL {seconds} SECOND");
@@ -4291,14 +4264,6 @@ fn spinner_bar(message: &str) -> ProgressBar {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn parse_duration_to_seconds_rejects_non_ascii_suffix_without_panicking() {
-        for input in ["1–", "1é", "1🙂"] {
-            let err = parse_duration_to_seconds(input).expect_err("invalid unicode suffix");
-            assert!(err.to_string().contains("invalid duration"));
-        }
-    }
 
     #[test]
     fn push_checkpoint_line_offset_advances_only_after_commit() {
