+++ src/utils/mod.rs
@@ -1,7 +1,9 @@
+mod duration;
 mod fs_atomic;
 mod git;
 mod plurals;
 
+pub use duration::parse_duration_to_seconds;
 pub use fs_atomic::write_text_atomic;
 pub use git::GitRepo;
 pub use plurals::pluralize;
